refactor merkle tree built and split db part
