+++ merkle/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod handler;
 pub mod merkle;
+pub mod db;
